0.949  2022/02/10  SNOOPYJC  Handle ${\(expr)} and @{[expr]} in strings, fix while in stmt modifier with global pattern, fix AUTODIE, uniform handling of TRACEBACK using cluck, fix system, fix empty signal function, implement hex and oct

+++ pyf/_truncate.py
@@ -3,12 +3,18 @@ def _truncate(fh, length):
     """Implementation of perl $fh->truncate method"""
     global OS_ERROR, TRACEBACK, AUTODIE
     try:
-        os.truncate(fh, length)
+        if hasattr(fh, 'truncate'):
+            fh.truncate(length)
+        else:
+            os.truncate(fh, length)
         return True
     except Exception as _e:
         OS_ERROR = str(_e)
         if TRACEBACK:
-            traceback.print_exc()
+            if isinstance(fh, str):
+                _cluck(f"truncate({fh}, {length}) failed: {OS_ERROR}",skip=2)
+            else:
+                _cluck(f"truncate to {length} failed: {OS_ERROR}",skip=2)
         if AUTODIE:
             raise
         return None
